Bug 461354: [GTK] Find a better solution for inserting elements into Table/Tree/List
Fetching the topIndex after the user scrolls works perfectly fine. The
use case that fails is when setTopIndex() is called followed by a call
to getTopIndex(). The bug is exhibited by setting and fetching the
topIndex programmatically (not through the UI).

The fix for this is to keep track of the topIndex using a variable. The
idea is to use GTK API to fetch the topIndex when the user has scrolled
with the UI, and use the topIndex variable when it has been set
programmatically.

This patch fixes 3 of the 8 failing test cases on CentOS. See bug
493357.

Change-Id: Id9b12cd85ce121e4ecdd41fe3b27e803f89dfaa4
Signed-off-by: Eric Williams <ericwill@redhat.com>