Bug 493015: [GTK3] Text widget with SINGLE style and custom background color: selection not visible
This bug affects Text widgets with a custom foreground/background color
set. The selection background and foreground colors are wrong, the
selection foreground color should be COLOR_LIST_SELECTION_TEXT and the
selection background color should be COLOR_LIST_SELECTION. For some
versions of GTK less than 3.20, improper or no colors were being set at
all.

The fix is to set COLOR_LIST_SELECTION_TEXT when setting the foreground
color of a widget, and to ensure COLOR_LIST_SELECTION is set as a
selected background for Text widgets. This is accomplished via CSS and
gtk_widget_override_color() (for GTK3.14 and below only).

Tested on GTK3.20, 3.18, 3.16, and 3.14. GTK2 is unaffected, no
additional AllNonBrowser JUnit tests fail on GTK3.

Change-Id: I740d3fddf757570b475cf0caa8253f659ac8141b
Signed-off-by: Eric Williams <ericwill@redhat.com>