Bug 493036: [GTK3] Tabs within an editor have bad background (EGerrit)
Using GtkNotebook {background: ... ;} as a selector for the background
color of a TabFolder results in the background color bleeding outside of
the area enclosed by the border. This results in weird looking colors.

The fix is to use the GtkNotebook.header selector instead.

Tested on GTK3.20, 3.18, 3.16, and 3.14. GTK2 is unaffected, no
additional AllNonBrowser JUnit tests fail on GTK3.

Change-Id: I05e80f54aea3c82a55407a94967f41f71cab4864
Signed-off-by: Eric Williams <ericwill@redhat.com>