Bug 487160: [GTK3] Using ScrolledComposite in a multi-page Wizard results in improper sizing of widgets
This bug is caused by getClientArea() returning the wrong size for a
child Composite, after setBounds() has been called on the parent.

In Control/Composite.setBounds(), the parent Composite/SwtFixed
container has its size adjusted by swt_fixed_resize(). In GTK this is
supposed to "trickle down" to the container's children and adjust their
sizes as well. GTK3 caches sizes, so setting a size on the parent and
fetching the child's size shortly thereafter might not always return the
updated size.

The fix is to allocate the size of the parent right away in
swt_fixed_resize(). This causes the sizing machinery to trigger a lot
earlier and update the size accordingly, prevent SWT from fetching the
wrong size and making a miscalculation.

Tested on GTK3.20, 3.18, 3.16, 3.14, and 2.24. No additional
AllNonBrowser JUnit test failures on GTK3 or GTK2.

Change-Id: I1cadd7a81de5d0f4fceb039a03ff5870992c2d19
Signed-off-by: Eric Williams <ericwill@redhat.com>