Bug 494958 - Add color tab to ControlExample
I have added a new "Color" tab into ControlExample that displays all
available colors to use in the format Name, Type, RGB, and the actual
color. The colors used are from SWT.java and are put into a hash table
with the id as its key and the name as its value.

I was able to get the RGB color codes by building a method called
getRGBColor(), which grabs the Color object using
Display.getSystemColor() and then building the string using the
getRed(), getBlue(), and getGreen() methods from the color object.

The colortab class extends from the Tab class, but a lot of the methods
added were from the Tables class. I also overloaded the
createControlGroup() method to create a custom Parameters group (located
on the right side of the example group). This allowed me to remove a lot
of unnecessary controls for my color table as they are not relevant.

Change-Id: I7f650ceada799dfe0252155efa18cf2ff78b2910
Signed-off-by: Ian Pun <ipun@redhat.com>