Bug 487757 ScrolledComposite is empty if it is hidden before showing
Deal with bug where for Gtk3.8+, ScrolledComposites appear empty or
internal widgets don't
get sized properly. When resizing shell, widgets appear or grow into
correct size.

The issue is that upon widget creation ZERO_HEIGHT/WIDTH is assigned to
widget's state (createWidget());
then if you call setVisible(false), setVisible(true) before a layout
operation,
in the setVisible(true) call, the widget is never shown. In Gtk3.8+
hidden widgets have a
size of 0x0, and as such allocation is not done correctly until you
manually re-size the parent.
To resolve the issue we check if widget is not HIDDEN on SWT side, but
is hidden on GTK side. If it
is, then show it. This correctly initializes layout of internal widgets.

Change-Id: Ifa07fa0845fb8d951a8fbbca7296c0387b0b946f
Signed-off-by: Lev Ufimtsev <lufimtse@redhat.com>