Bug 495909 - [GTK3.20+] test_setTopItemLorg_eclipse_swt_widgets_TreeItem() fails
Referring to bug Bug 461354, I used a topIndex variable to keep track of
the top item. Check to see if there has been any scrolling changes done
by the user, and if not, return the topIndex variable. We can keep track
of scrolling changes by using a cachedAdjustment variable with a
currentAdjustment variable grabbed with GTK functions.

This optimization has also been adapted on the Table widget.

Tested on GTK 2, 3.14, .16, .18, and .20.

Change-Id: Ica1bda3c3d23d5cd27b8c4e29963c0f097d778d3
Signed-off-by: Ian Pun <ipun@redhat.com>