Bug 496153 - Remove useless non-X11 branch in Control.setEnabled
There is code guarded with if (!OS.GDK_WINDOWING_X11 ()). Such a check
is useless nowadays because on GTK3 it will check whether the x11
backend has been compiled which is always true as GTK3 supports multiple
backends compiled in (difference with GTK2). And on GTK2 there are no
non X11 backends working.
Helps when checking Wayland/Broadway issues to not be confused by such
potential branches.

Change-Id: Ifa20cf71c0c5fbd1da3544c18c005d224ae7ca2c
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>