Bug 498829 - Load webkit2 so file if webkit 1.x is not available
When libwebkitgtk-3.0.so.0 can't be dlopen, try to dlopen libwebkit2gtk
to run in webkit2 mode.
As the SWT_WEBKIT2 env variable is not set in this case try calling
webkit_get_major_version which is Webkit2 function and due to dlsym
failing it will always return 0. If result is not 0 libwebkit2gtk so
file was loaded as fallback.


Change-Id: I5a7f66d2b122a5e2368b20bd6ed76b04e654643c
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>