Bug 498829 - Load webkit2 so file if webkit 1.x is not available
** R4_6 Backport **
When libwebkitgtk-3.0.so.0 can't be dlopen, try to dlopen libwebkit2gtk
to run in webkit2 mode.
As the SWT_WEBKIT2 env variable is not set in this case try calling
webkit_get_major_version which is Webkit2 function and due to dlsym
failing it will always return 0. If result is not 0 libwebkit2gtk so
file was loaded as fallback.


Change-Id: Ic7dd75f978c0618987cab859e4f14ce5efc40e77
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>