Bug 499380: [GTK3.16+] ToolBar loses background color if setting background and foreground consecutively
Since ToolItems are not children of Control, they do not carry the
cssBackground and cssForeground caching properties. This means caching
the foreground of a ToolBar (from each of its ToolItems) by overriding
setForegroundColor() in ToolBar. This only affects GTK3.16 onwards.

Included in this patch is a test case to detect this issue. Tested on
all versions of GTK3, and GTK2.24. No additional AllNonBrowser JUnit
tests fail.

Change-Id: Iebf4c9d1aa13a7d0a3cfc0dccfacfa24998f465f
Signed-off-by: Eric Williams <ericwill@redhat.com>