Bug 94549 - New Java class wizard "Interfaces" table displays junk
Using Billy Biggs' patch to fix positioning of the text editing area.
The patch takes account for the cell size in order to get the correct
height for the tree for the text editing widget to spawn correctly.
Jface seems to call editElement() even when the table does not have
focus, causing sizing issues.

Change-Id: I075d634c73c1b92c28f8f4854c5e35a8ce6a0acd
Signed-off-by: Ian Pun <ipun@redhat.com>