Bug 500363: Native auto generated files should mention that they are auto generated
When doing work with native files, it's not clear as to which files are
auto generated and which files are written manually. Especially when you
are just starting to work with native C code.

I added a method to JNIGenerator.java that inserts a note to every
generated file about the file being auto-generated.

[Update Patchset 3]
- The updated patchset updates contains auto-generated files for all
platforms.
- Fixed typo and added name of generator class.

Change-Id: I7fdac972eb2cb829316d0e970601ed4c87e6008b
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=500363
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>