Bug 500781 - Allow widgets to define which handle is their css one
Some SWT widgets don't map 1:1 with GTK widgets thus their handle is not
the one one would like to apply the css to. Thus cssHandle method is
introduced and Label is the first to override it by pointing to its
labelHandle instead. 
Adjusted Widget.setData to apply the gtk css to the return value of
cssHandle.

Change-Id: I721b40442f043506fceeeea88cd2f0ef44229bb1
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>