Bug 501855: [GTK] Add script to check which dynamic functions are deprecated
This script (when run) compares all dynamic functions from os_custom.h
with a list of deprecated GTK3 symbols. The list of GTK3 symbols is
fetched from the GTK website in order to provide up to date information.

Most of these functions will already have been handled, but some become
deprecated over time. We are unable to catch these cases with the
compiler since these functions are not compiled. For these cases, this
script is useful.

Usage: 

Running without any options prints a list of deprecated dynamic
functions to console. 

Running with the "-f" option will write the list of dynamic deprecated
functions to a file calld "dynamic_deprecated_functions.txt", which is
stored in the same directory as the script.

Change-Id: Iac32086b8e1225545a92a05b68a8cfa84542b76e
Signed-off-by: Eric Williams <ericwill@redhat.com>