Bug 431319 - Run SWT tests with plain surefire (not Tycho/OSGi one)
This configures the tests to run with the plain Java
maven-surefire-plugin. To do so, it requires a feature listing all
platform-specific fragments that we add as a dependency, and to make
target resolution platform-specific. Then Tycho generates a good
classpath that's consumed by maven-surefire-plugin.
Printing is disabled as the tests are failing with tycho.

Change-Id: Ibba7c64757fb8ec71aff04c850e35309440e3caa
Signed-off-by: Mickael Istria <mistria@redhat.com>
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>