Bug 507038 - Stop calling gtk_tree_selection_select_iter in List/Table.selectFocusIndex
From the comment: 
For some reason, when an event loop is run from within a key pressed
handler and a dialog is displayed that contains a GtkTreeView,
gtk_tree_view_set_cursor() does not set the cursor or select the item.
The fix is to select the item with gtk_tree_selection_select_iter() as
well. NOTE: This happens in GTK 2.2.1 and is fixed in GTK 2.2.4.

Clearly not needed with Gtk 2.24 and creates useless overhead.

Change-Id: I8a252dac46427c07d838f37258a192b5d927c31c
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>