Bug 508692 - Shells has wrong size under wayland
Reason for that is that is the usage of GtkWindow allocation for the
shell size which would fail with CSD (Client Side Decoration) as window
decorations, shadows and etc. are part of the window allocation. It
exposes on Wayland and Broadway backends as they have no server side
decorations at all.

Change-Id: I2a17a2cd66f486c19c0ed0cb2055aa52a3c7f7e4
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>