Bug 508692 - Shells has wrong size under wayland
Reason for that is that is the usage of GtkWindow allocation for the
shell size which would fail with CSD (Client Side Decoration) as window
decorations, shadows and etc. are part of the window allocation. It
exposes on Wayland and Broadway backends as they have no server side
decorations at all.

Reverted the patch to include conditional GTK3 checks as removing it
caused Bug 509596.

Change-Id: If7d3f45739bfd1849ff53eeee7c13f7024a05385
Signed-off-by: Ian Pun <ipun@redhat.com>