Bug 509411 - [Webkit 1] Crash during Test_*_Browser evaluate() tests (2)
1) 
Skip newly added Browser.evaluate() tests (#505591) for webkit1 because
webkit1 is not stable with these and hudson consistently crashes.
Note: These tests do not fail on Cocoa/Win32 and Webkit2.

2) 
Added a few notes to few other tests that cause sporadic crashes on
webkit1 due to use of context_iteration (that also don't occur on
webkit2). See bug for details.

3)
Also changing setup/teardown of legacy tests so that they don't run 
for tests that are skipped on hudson. #509658

4) Tidy/refactor

Change-Id: I35a398ad56f36e7a618537cb6eca492c8ef7f1ac
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>