Bug 509627 - Do not manually draw CURSOR_APPSTARTING
Instead use the named cursor which will fit in the overall cursor theme
(if such is available) and not be the ugly and ancient looking custom
drawn cursor.
But let it fallback to the old X cursors loading if the named one fails
so old systems and/or Xvnc continue to work properly.

Change-Id: Ib324794c9fef4d2f14cfe557bfbc19c92edce5c3
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>