Bug 500013 - [Wayland] Keyboard Left and Right Arrow keys are very slow or unresponsive
The Caret drawing mechanism was originally implemented to be done using
a generated cairo from parent window using gdk_cairo_create(). This is
now deprecated and not working in Wayland. Drawing of the caret is now
handled in a draw signal which is triggered by a queue_draw() call.

Fixed a race condition issue caused by a variable introduced called
isDrawing. Removed it so cursor as it was not necessary and will not
cause the caret to disappear anymore. 

Change-Id: I7be19f1982a18145c6192b266459ea9d3472d4fb
Signed-off-by: Ian Pun <ipun@redhat.com>