Bug 511355 - Table: background color of sort column became a darker gray
Change the algorithm for the sort columns of Table and Tree, so that for
the most common use case (white background) the result is the same as
before, but also supporting black background. The 'getDifferentColor'
algorithm stays in place for highlighting buttons as well as Table and
Tree headers, as this produces a higher contrast compared to the sort
column color algorithm.

Change-Id: I93d09203786cb66bbb1eefcf464f5aff0e4d2f45
Signed-off-by: Conrad Groth <info@conrad-groth.de>