Bug 503431 - [wayland] Mouse button release event not triggering from gdk_event_get() causing DND timeout
- Reworked DND logic to be done through the use of GTKGestures, which
are created in the constructor of a Control.
- Moved DragDetect() call to be used in the motion notify event callback
in order to get GTKGesture mouse position offsets correctly. 
- Changed labels to use gtk_event_box instead of box in order to have
them output correct signals when being dragged. GtkBox is currently
nested inside a GTk Event Box in order to preserve images.
- Reworked tree/list/table so that drag and drop with multiple items
selected work. Selection method is now called in a conditional call in
release_event() as opposed to press_event(). 
- added dragEndReleaseSelection() in Control to be used by dragSource in
order to fix issue where release events are not called during DnD in
Wayland. The fix was to have the logic in the release event from
multiselection widgets in the connected signal DragEnd as well.

Change-Id: I3c8f2e90069c874cd831d44a5ca8b23fd70558ad
Signed-off-by: Ian Pun <ipun@redhat.com>