Bug 512159 - [CTabFolder] Several getter methods in CTabFolder do not use checkWidget()
Not sure why these methods do not use checkWidget. The last commit
message for Silenio is: "restore HEAD after accidental deletion by error
in automated build script" which is not very informative.

Leo:
While reviewing, I found a bunch of other places that had checkWidget()
commented out.
The original commit that added them has an empty log message
and is not linked to bugzilla: 
    6a87ceb8219d8d816e8c34a66f6f56af4e43ead7
	Author: Veronika Irvine <veronika>
	Date:   Fri Feb 15 13:55:04 2002 +0000
    *** empty log message ***
I can only presume that it might have been done so that fields
can be accessed on the same thread without having to get the UI
thread for convienience. But for sake of consistency, we should
have these checks in all public methods.

Change-Id: I719fb7edf56fd6bface0eaf1b501856af0996b74
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>