Bug 511874 temporarily disable test_execute_and_closeListener osx/win32
This test seems to be inconsistent on Cocoa and Win32.
It may be an issue with the test itself.

For now, I'm making this test only run on webkit*

I will setup a Cocoa/Win32 box to investigate this after vacation in 2
weeks. For now to prevent build log errors, I'll turn the test off.

Change-Id: I40e2ad2ff8afcfe0efa5e47f080e46312aee0f8f
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>