Bug 508786 - [wayland] Shell.setMinimized(true) doesn't work on wayland
setMinimized call now checks to see if the shell is visible or not and
sets it to true if it isn't before calling gtk_window_iconify().

Can be tested with Snippet27 in both X11 and Wayland. Using Gnome, go to
gnome-tweak-tool:extensions and turn on window list. The snippet should
run but be greyed out in the window list, indicating it is minimized.

Change-Id: I739a5e4db371b7b18aa9a7073733e38ad11766d5
Signed-off-by: Ian Pun <ipun@redhat.com>