Bug 511874 : Fix jUnit test, Browser on Win32 has one child.
On Win32, Browser has one child, an OleFrame. 
This causes breakage in test_getChildren() if the shell's composite
is set to be the browser.
Solution: On Win32's jUnit, override test_getChildren and treat it in
a special way.
Note: test_execute_and_closeListener() has no issues if setWidget() is
used in constructor.

Patch's jUnit tested on Gtk3 webkit2, OSX, Win32.

Change-Id: I15e09ec4af3d459ff937e979fe4a030f7656a272
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>