Bug 513492: [webkit2] Port VisibilityWindow event details
Implemented/ported Visibility event for webkit2.

Before Snippet270, when you click on "try it", nothing happened on 
webkit2. Now a popup correctly opens.

This also *improves* Browser5 test, in so far as it almost passes.
(Currently fails due to another cause thou, to be fixed in another
patch).

Architecture note: I copied GdkRectangle from OS instead of linking to
it, because otherwise all of gdk/gtk dependencies are bought into
webkit.
It didn't seem that all of those dependencies are needed just for
GdkRectangle.

Change-Id: I3e8ff3d007d6eabe183b9734eb6adc2020e4e8fc
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=513492
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>