Bug 514419 - [regression][GTK3] DND is broken in staging view
Fixed issue regarding DnD being broken in staging view. The issue was
that the moment you left the tree path (outside of any elements), the
drag detection call would not be called, and would return false. I
adjusted it so that it would check using
gtk_gesture_drag_get_start_point to see if the drag event started inside
within the tree path, but continue to check for DnD detection using the
starting x/y coordinates to check threshold. This will also improve
overall performance as there will be less GTK calls needed, as the
dragGesture needs to be active in order for any further gtk threshold
calls are to be made. 

Tested on GTK2, GTK3.22, GTK3.14, GTK3.22 on wayland.

Change-Id: Ie9a07af418e9eee987e3f02a8fc5f9044c18d0f7
Signed-off-by: Ian Pun <ipun@redhat.com>