Bug 515035 - [Wayland] DnD on text causes null pointer exception
Fixed issue that causes a null pointer exception called on GTKText (SWT
Text) widget. Reasoning behind the changes is that in GTK3.14 there was
a change to how GTK handles DnD automatically for GTKText, causing drag
detection from SWT to fight over GTK. In the case that GTK were to win
the DnD, SWT was not signaled about that event, thus causing a null
pointer when the target item is copied/moved as the target item was
never instantiated.

From my investigation, the Text field was broken way before my patch,
and seem to have been caused by GTK3.14 unnoticed.

Change-Id: I23d102d49c97f9c8b138d355cd01ee49f3cfc19a
Signed-off-by: Ian Pun <ipun@redhat.com>