Bug 515035 - [Wayland] DnD on text causes null pointer exception
Fixed a potential issue that would cause both SWT and GTK to fight over
who sets the drag first. Patch will make sure all DND will be done over
at GTKText level and then referenced back to SWT. 

This can be tested through DnDExample -> Text. Previously it would fail
on certain occasions (if the drag effect is a text vs. a file icon). The
patch now forces it to be through GTK3 DnD which is the text icon, and
in turn, will signal back to SWT that a drag is occurring. Text is set
to SWT.MULTI in DnDExample so I swapped it to SWT.SINGLE as well for
testing as they call either GTKText or GTKEntry respectively.

Also tested through AllBrowserTests on GTK2, 3.14,3.16,3.18,3.20,
Wayland with no additional errors. 

Change-Id: Id3bfef03f0ef84e6e341782124ae4bda40f5984e
Signed-off-by: Ian Pun <ipun@redhat.com>