Bug 515155 - Avoid memcpy@GLIBC_2.14 for old Linux compatibility
GCC by default optimizes memmove() calls to use __builtin_memmove(),
which in turn uses memcpy() from the system GLIBC if source and target
regions are known to not overlap. When compiling on a modern Linux
host, this leads to pulling memcpy@GLIBC_2.14, which is not available
on older Linux hosts.

In order to be compatible with RHEL6/Centos6 level systems, we avoid
the memcpy@GLIBC_2.14 symbol by -fno-builtin-memmove when compiling.

This is only done for gtk2; we do not care about gtk3, since systems
that have gtk3 are supposed to also have memcpy@GLIBC_2.14.

Change-Id: I50fec4cbb7d662f27a28693543b2f2024a8ea375
Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>