Bug 515610 - [GTK] SWT generates incorrect fall-through for memmove()
An incorrect fall-through in generated C code for memmove() led to
producing code which could memmove() to NULL pointer destination
or from a NULL pointer source.

Even in case this was just dead code and could never execute, it led
to an unnecessary inefficiency in the default case (an extra jmp
instruction to jump over the dead code).

This is resolved by a better generator, which generates an earlier jump
to fail: thus avoiding the undesired fall-through.

Patchset 4: (Leo)
- Win32's OS.java doesn't have memmove, only MoveMemory, so Win32 not
affected.
- Adding Cocoa's resulting os.c of this patch.
 (I launched a child eclipse with this patch to have newest SWT tools,
  then set .classpath_cocoa -> .classpath and modified OS.java for
cocoa)

Change-Id: I9f9153ae3f41f4aea3a9f9439e92a892224b1475
Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>