Bug 514483 - [wayland] Launchbar filter closes when clicked
GetCursorLocation() was returning wrong coordinates. In Wayland
applications do not share a global coordinate, instead it is relative to
the parent window. The position of the cursor was returning relative to
the popup window and not of the parent window, causing the LaunchBar
logic to think that we are clicking outside of the popup window, thus
closing it. Fixed by using the relative position as an offset and adding
it to the position of the popup window, which will return the correct
coordinates.

Testing this can be done through snippetA in
https://bugs.eclipse.org/bugs/show_bug.cgi?id=514483 . 

Tested with Junits on Wayland with no additional failures/errors.

Change-Id: I9bfb2052e59de38a93f75cc624759b41885ae2a6
Signed-off-by: Ian Pun <ipun@redhat.com>