Bug 97863 - [consistency] dragging TableColumns fires too many control events
There is now a new check on minimumResizeSignal which will only have
tableColumn check for move/resize if the widget isn't allocated to be a
1x1 size at position 0,0. This is to fix the issue of dragging
tabcolumns causing unreliable control event spam.

This can be tested through the snippet in Bug 97863 Comment 1.

Tested with AllNoneBrowserTests for GTK2, GTk3.22, 3.18, 3.14 with no
additional errors.

Change-Id: I06061df7c256644e2a9a8587d982f462796ca24c
Signed-off-by: Ian Pun <ipun@redhat.com>