Bug 517633 - Strange text replacement in Spinner
Added additional conditional barrier as there was postfix 0s that were
lost during the conversion between string -> double. This will fix the
issue with spinners losing postfix 0s and also keep selection signals
working correctly. This can be tested through the snippet in the bug
517633 attachment.

Tested with AllnonBrowserTests gtk2 to gtk3.22 and Wayland. 

Change-Id: Ia6bd3dedc49f9c43d28953dff9ae113d7520d50b
Signed-off-by: Ian Pun <ipun@redhat.com>