Bug 494565: [GTK] Snippet243 doesn't work because Text#insert(String) modifies topIndex
Methods that insert or select text should not modify the topIndex of the
viewer.

To avoid this issue we calculate the visible area, positions of the
topIndex, and the insertion/selection points. If the insertion/selection
points are outside the visible area, then scroll to them. Otherwise do
nothing, which preserves the topIndex.

Tested on GTK3.22, 3.20, 3.18, 3.16, 3.14, 3.10, and 2.24. No additional
AllNonBrowser JUnit test failures occur.

Change-Id: Ib020473a5a31d535fb0b1b84f8b13e087fa16c62
Signed-off-by: Eric Williams <ericwill@redhat.com>