Bug 486068: [GTK3.20+] Allocation warnings printed in error console
This fix targets the Group widget.

GTK3.20+ calculates allocation sizes using GtkCssNode elements which we
cannot access. The fix is to calculate the following:

to-be-allocated width/height - (difference between the preferred
width/height and the to-be-allocated width/height)

If the above number is below 0, allocate the preferred width/height.

Tested on GTK3.22, 3.20, 3.18, 3.16, 3.10, and 2.24. No additional
AllNonBrowser JUnit test failures occur.

Change-Id: Ie8e521aa35d3c411411a5ee872ea79c407dd1dce
Signed-off-by: Eric Williams <ericwill@redhat.com>