Bug 519124 – Add SWT variable to print versions of underlying libraries
Adding code to print gtk/webkit versions if SWT_LIB_VERSIONS is
specified.
This is helpful for bug report investigations.

Note, SWT_LIB_ prefix is for easier grep-ing of console output in the
situation if there is a lot of console output.

Console output is like:
SWT_LIB_Gtk:3.22.15
SWT_LIB_WebkitGtk:2.16.0  (Version >=2.5 is Webkit2)

Change-Id: I753e006d8f0464ca51035400327c43550a3f3c31
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>