Bug 519321: [GTK3] Preferences window sometimes has white background
In Control.setBackgroundGdkRGBA(): if the incoming GdkRGBA is null, then
the background color needs to be reset to COLOR_WIDGET_BACKGROUND. 

On GTK3.16+ there is an automatic check with GTK CSS to set incoming
GdkRGBA's as COLOR_WIDGET_BACKGROUND if they are null. Setting a null
GdkRGBA using GtkStyleContext machinery resets the current color to its
previous value, which now (after bug 477950) could sometimes be
transparent. This is what was causing the bug. The fix is to set
incoming GdkRGBA's that are null to COLOR_WIDGET_BACKGROUND. Of course
not every widget uses COLOR_WIDGET_BACKGROUND as the default, so it's
been overriden in those places.

This ensures GTK3.14 and below still works correctly when
setBackground(null) is called.

Change-Id: Ib635be9ed61edf8c8e602fe3e27dd8c418210b12
Signed-off-by: Eric Williams <ericwill@redhat.com>