Bug 518379 – Allow platform specific JUnit tests(gtk tests not on cocoa)
Currently swt gtk only tests are marked to compile/build on cocoa,
because Cocoa is a 'unix' family type (as is Linux).

The correct solution is to add "<name>Linux</name>" to activation
profile. This way the Unix profile is only launched on Linux and not on
Cocoa.

Found this in:
http://maven.40175.n5.nabble.com/Profile-activation-for-mac-and-linux-td3263043.html
https://stackoverflow.com/questions/38465711/choose-maven-profile-from-os-family/44975462#44975462

As a note, family names are defined here:
https://github.com/sonatype/plexus-utils/blob/f2beca21c75084986b49b3ab7b5f0f988021dcea/src/main/java/org/codehaus/plexus/util/Os.java#L72

Verified on:
- Fedora/Ubuntu/CentOS (gtk tests included)
- Cocoa (gtk tests not included)

Change-Id: I51dd2cd9de9e6efc8c348d1eec083325c2774b2a
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>