Bug 519916 pom.xml doesn't have to specify webkit as default browser anymore
In tests/../pom.xml, we hard-code the default browser to be webkit. This
causes maven tests related to Browser to fail on Windows as it's trying
to use webkit instead of I.e, which by default is not installed.

I.e:
<argLine> .. -Dorg.eclipse.swt.browser.DefaultType=webkit ... </argLine>

Since XUL Runner is removed from codebase, hard-coding webkit is no
longer needed. As of recently:
- On Linux, webkit is default 
- On Cocoa, webkit is default
- On Win,   I.E is default

Thus the hard-coding can be safely removed without breakage.


Change-Id: I815d3ac45f4c957f0d6f79066d27f2a1d98b2273
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>