Bug 517487 - [GTK3][Regression] Caret disappears when moving/editing
Fixed issue of caret dissapering in wayland and X11. The reoccuring
blink proc will be reset everytime setVisible is used so that blink rate
can start the moment it is moved. Using setBlink() will now make sure
that gtk_draw() signals that are not being sent from the blink proc are
ignored. This was caused by the spam of draw signals from the scrollbar
being visible.

Tested on both X11 and Wayland for AllNonBrowserTests junits GTK 3.22,
3.18, and 2.

Change-Id: Id0ab900300a56f64957db91c74c22a32f527332d
Signed-off-by: Ian Pun <ipun@redhat.com>