Bug 517264 Refactor old Browser* tests -- Skip test_setUrl_remote on bad connection.
test_setUrl_remote() requires a live Internet connection.
This test sometimes times out when a build server happens to have a
bad Internet connection.

This causes noise in the test logs. 

It's difficult to ensure live connectivity is always present,
so instead this test will be skipped if there is no internet present.
Note, setUrl(..) is still tested against local files even if
an internet connection is not available.

Change-Id: I098c1ecf3707c23260440fa0cf9081805198a627
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>