Bug 508129 - [10.11]PaintEvent.GC.copyArea () crashes Eclipse Neon on OSX 10.11 and above
When a GC is created with a control as the Drawable and GC.copyArea() is
called, a SWT.Paint event is sent unexpectedly. This happens because
NSView.cacheDisplayInRect() calls drawRect() which causes a SWT.Paint
event to be sent. The fix is to ignore this paint event and prevent SWT
from sending it to the clients. 

This is handled by creating a custom callback for cacheDisplayInRect,
setting the ignorePaint flag and then calling the Cocoa implementation
of cacheDisplayInRect. The ignorePaint flag is reset after the call.
	

Change-Id: I4c2e2986b83868082295ea187b543376bf4c8a7e