Bug 519089:[Java 9] Automatic window tabbing should be disallowed
In macOS 10.12, a new system preference "prefer tabs when opening
documents" has been added which causes automatic tabbing of windows in
Eclipse. This is by default enabled to work when application is in
fullscreen mode and can be modified to work always.

Disable automatic window tabbing, but setting the
NSWindow.allowsAutomaticWindowTabbing property to false.

Change-Id: I7f22249d92dabe2597147b9089c42f79b68ac6c4