Bug 522733 - Crash on close with webkitgtk 2.18 (Workaround)
(Backport to 4.7.2)

As of Webkitgtk 2.18, webkitgtk2 crashes if the first instance of
webview is not referenced when JVM shuts down.
There is a exit handler that tries to dereference the first instance
[which if not referenced] leads to a crash. This workaround would
benefit from deeper investigation (find root cause etc...).

With that said, this patch only adds a single reference to a single
instance of webview, which is cleaned up on JVM shutdown. I believe
this doesn't leave a memory leak behind.

This bug might potentially be caused by some race condition as well.
e.g webkitgtk still doing something while jvm is shutting down etc...

Tests:
- Browser JUnit tests on Webkit1 & Webkit2.

Change-Id: I414af620cb82a745f1c3bd3c4c67ce53ffe4c320
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>