Bug 525763: [GTK3] [dark theme] Vertical bar to the left of line numbers is not dark
This fix targets an issue on the GTK3 dark theme where the side rulers
have a white background.

Part of the fix was to revert some of the background color machinery to
pre color port functionality. Some fixes in bug 477950 were made to
prevent another issue from happening where ToolBars that are children of
CTabFolders have a light background color. The proper fix for that issue
is to use !getBackground().equals(color) instead of always setting the
background color.

Tested on GTK3.8 to 3.22, and GTK2.24. Tested under normal, dark, and
classic themes.

Change-Id: Ie70c050c954c443bf2445407c49fb9cb7d84c1e3
Signed-off-by: Eric Williams <ericwill@redhat.com>