Bug 525763: [GTK3] [dark theme] Vertical bar to the left of line numbers is not dark
This fix targets an issue on the GTK3 dark theme where the side rulers
have a white background.

Part of the fix was to revert some of the background color machinery to
pre color port functionality. Some fixes in bug 477950 were made to
prevent another issue from happening where ToolBars that are children of
CTabFolders have a light background color. The proper fix for that issue
is to use !getBackground().equals(color) instead of always setting the
background color.

In addition, the logic for setting a background color when the
PARENT_BACKGROUND flag is set, has changed. If findBackgroundControl()
returns null, set the current control's background color to transparent,
to make sure the color actually gets inherited.

Tested on GTK3.8 to 3.22, and GTK2.24. Tested under normal, dark, and
classic themes.

Change-Id: Id0f6cff24e72e0ed4550b066720dd87d006d9ea9
Signed-off-by: Eric Williams <ericwill@redhat.com>