Bug 470031: [GTK3] Some viewparts became inaccessible in 4.5
AtkTable interface implementation. Not that CTable (which makes use of
these interfaces) uses CTableItems for its children. Currently
accessibility support for children of emulated widgets is limited.
Because of this, snippets like AccessibleTableExample are not working
100% at the moment.

This patch also adds Javadocs for AtkSelection functions in
AccessibleObject, since they were not added earlier.

Change-Id: Icc25bf41b92a0aacb615ea79bf97f02738c31359
Signed-off-by: Eric Williams <ericwill@redhat.com>