Bug 527536: [GTK3] NoSuchMethodError: gObjectClass_finalize
Call AccessibleObject.gObjectClass_finalize only if the
SwtFixedAccessible instance has a matching Java accessible object. Chain
up to the parent in os_custom.c for all cases anyways.

This patch guards the method ID lookup as well, and returns 0 if there
is an issue. This way it should be impossible to get a NoSuchMethodError
thrown.

Change-Id: I285666e07e66adc2c677964b3d8d4c4bde29b6b2
Signed-off-by: Eric Williams <ericwill@redhat.com>