Bug 527737 [Webkit2] regression, execute() script should not be wrapped into a function
I've introduced a regression in the past. execute() should not wrap it's
script into a function like evaluate(), because it affects
the scope of the function that is being executed.

All Test*Browser jUnits continue to work as before. This fixes
snippet326 in that it now displays the proper dialogue as expected.

Change-Id: I92fbe59b655d057c6e022dd03850e5344aadb407
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>