Bug 527738 [Webkit2] implement setUrl(...POST..) support. (snippet 330)
When setURL contains POST data, bypass the Webkit2 API and use the Java
API (HttpURLConnection) to perform the request. Set the URL globally so
that when calling setText with the response string, any relative paths
can be resolved using the base URL.

This implementation does not currently support Webkit configured proxies
and there may be issues with detecting various signals/listeners that
operate on request/response objects given that these are handled by the
Java side.

Tested against Snippet330.

---
Leo's modifications: Patchest 9 notes:
- Added g_bytes_unref() to prevent memory leakage.
- Making g_bytes_* dynamic otherwise we'll get jvm crash on gtk2.
  (g_byte* introduced in glib 2.32, but gtk2 only has glib 2.26(ish)).
- Move data type conversion logic into one place. (html converted where
it's used)
- Reduced size of if statment, it was scary.
- Add scope to code for clarity.
- Don't touch this.postData & this.headers, they're only used by
webkit1.
  Webkit2 shouldn't touch what it doesn't use.
- Added if (Webkit1) guards around variables only used by webkit1.
- Refactoring.
- Avoid passing null value to webkit_web_view_load_bytes(.) if
connection is not HTML.
- Instead of passing empty array to encoding type, pass array with only
a single 0.
- Modified Snippet330.java to include some more examples for better
testing and learning.
- Verified:
 - against (modified) Snippet330
 - jUnit tests. (w1/w2)

...

Patchset 15:
- Added counter support to setText(..) as a long setURI() could override
  a setText() call.

Change-Id: Ic93747083af6af9933af3ca52149a1fb6e894d2c
Signed-off-by: Roland Grunberg <rgrunber@redhat.com>
Signed-off-by: Leo Ufimtsev <lufimtse@redhat.com>